feat: create Purchase class
